+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Add conditional Array `RbObjectConvertible` conformance.
 * Add conditional Dictionary `RbObjectConvertible` conformance.
 * Add `RbThread` utilities.
+* Add conditional Range family `RbObjectConvertible` conformance.
+* Add `RbObjectCollection` to use collection protocols with Ruby.
 
 ##### Bug Fixes